refactored code to use a single routes.js file passing in app and a resolver file
